restore visual changes, move example to visual tests
