+++ packages/fluentui/docs/src/examples/components/Button/Types/ButtonExample.shorthand.tsx
@@ -1,12 +1,6 @@
 import * as React from 'react';
 import { Button } from '@fluentui/react-northstar';
 
-const ButtonExample = () => (
-  <>
-    <Button content="Click here" design={{ border: '1px solid blue' }} />
-    <Button content="Click here" design={{ border: '1px solid blue' }} />
-    <Button content="Click here" design={{ border: '1px solid red' }} />
-  </>
-);
+const ButtonExample = () => <Button content="Click here" />;
 
 export default ButtonExample;
