+++ packages/fluentui/docs/src/examples/components/Button/Visual/index.tsx
@@ -6,6 +6,7 @@ import NonPublicSection from '../../../../components/ComponentDoc/NonPublicSecti
 const Types = () => (
   <NonPublicSection title="Visual tests">
     <ComponentExample examplePath="components/Button/Visual/ButtonExampleCompose" />
+    <ComponentExample examplePath="components/Button/Visual/ButtonExampleDesign" />
   </NonPublicSection>
 );
